backend-one/requirements.txt
Flask
pyrebase4
requests
python-dotenv
flask-cors